refactor: redo validation in the project. Annotations should be responsible for validating data in the project, the presence of a record in the database should be checked using separate methods in the DAO. Add an annotation that verifies the correctness of the release date of the movie

+++ src/main/java/ru/yandex/practicum/filmorate/service/DirectorService.java
@@ -3,6 +3,7 @@ package ru.yandex.practicum.filmorate.service;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import ru.yandex.practicum.filmorate.exception.NotFoundException;
+import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.Director;
 import ru.yandex.practicum.filmorate.storage.DirectorStorage;
 
@@ -26,8 +27,15 @@ public class DirectorService {
     }
 
     public Director updateDirector(Director director) {
+        Integer directorId = director.getId();
+        if (directorId == null) {
+            throw new ValidationException("Требуется корректный id параметр");
+        } else if (!directorStorage.existsById(directorId)) {
+            throw new NotFoundException("Режиссёр с id " + directorId + " не найден");
+        }
+
         directorStorage.update(director);
-        Director updatedDirector = directorStorage.findById(director.getId()).orElse(null);
+        Director updatedDirector = directorStorage.findById(directorId).orElse(null);
 
         log.info("Обновлён режиссёр: {}", updatedDirector);
         return updatedDirector;
@@ -47,6 +55,10 @@ public class DirectorService {
     }
 
     public void deleteDirector(Integer directorId) {
+        if (!directorStorage.existsById(directorId)) {
+            throw new NotFoundException("Режиссёр с id " + directorId + " не найден");
+        }
+
         log.info("Режиссёр с id {} удалён.", directorId);
         directorStorage.delete(directorId);
     }
